pyproject.toml
[project]
name = "nexusmods-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "playwright>=1.58.0",
    "pynxm>=0.1.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
]
